moved integration test to corresponding test packages instead of top level
